create new jmockit package and move files
